requirements.txt
networkx>=3.3
torch>=2.3.1
torch-geometric>=2.5.3
spacy>=3.7.5
pygithub>=2.3.0
streamlit>=1.37.1
fastapi>=0.112.0
uvicorn>=0.30.3
difflib
sqlite3  # Built-in, but noted for clarity
pytest>=8.3.2
vaderSentiment>=3.3.2  # For sentiment analysis
pylint>=3.2.6  # For code smell detection